feat : BaseAuth 생성 및 판매자&소비자 구현체 분리
