+++ src/main/java/com/project/member/controller/MemberController.java
@@ -8,10 +8,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.project.auth.service.AuthService;
 import com.project.member.dto.request.SignInRequest;
 import com.project.member.dto.request.SignUpRequest;
 import com.project.member.dto.response.TokenResponse;
+import com.project.member.service.MemberService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -19,11 +19,11 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 public class MemberController {
-	private final AuthService authService;
+	private final MemberService memberService;
 
 	@PostMapping("sign-up")
 	public ResponseEntity<Void> signUp(@Valid @RequestBody SignUpRequest signUpRequest) {
-		Long id = authService.signUp(signUpRequest);
+		Long id = memberService.signUp(signUpRequest);
 		URI uri = ServletUriComponentsBuilder.fromCurrentContextPath()
 			.path("/{id}")
 			.buildAndExpand(id).toUri();
@@ -33,7 +33,7 @@ public class MemberController {
 
 	@PostMapping("sign-in")
 	public ResponseEntity<TokenResponse> signIn(@Valid @RequestBody SignInRequest signInRequest) {
-		TokenResponse response = authService.signIn(signInRequest);
+		TokenResponse response = memberService.signIn(signInRequest);
 		return ResponseEntity.ok(response);
 	}
 }
