+++ src/main/java/com/project/member/domain/Member.java
@@ -2,7 +2,8 @@ package com.project.member.domain;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import jakarta.persistence.Column;
+import com.project.auth.domain.BaseAuth;
+
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -15,26 +16,18 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
+public class Member extends BaseAuth {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@Column(unique = true)
-	private Email email;
-
 	@Embedded
 	private Nickname nickname;
 
-	@Embedded
-	private Password password;
-
 	private Member(Email email, Nickname nickname, Password password) {
-		this.email = email;
+		super(email, password);
 		this.nickname = nickname;
-		this.password = password;
 	}
-
 	public static Member of(String email, String nickname, String password, PasswordEncoder passwordEncoder) {
 		return new Member(
 			Email.from(email),
@@ -43,20 +36,8 @@ public class Member {
 		);
 	}
 
-	public String getEmail() {
-		return this.email.getEmail();
-	}
-
 	public String getNickname() {
 		return this.nickname.getNickname();
-	}
-
-	public String getPassword() {
-		return password.getPassword();
-	}
-
-	public void checkPassword(String rawPassword, PasswordEncoder passwordEncoder) {
-		password.checkPassword(rawPassword, passwordEncoder);
 	}
 
 }
