+++ src/main/java/com/project/seller/SellerConstant.java
@@ -5,7 +5,6 @@ import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class SellerConstant {
-	//TODO int로 할 지 고민, notBlank, notNull도 이에 따라 달라짐
 	public static final String TAXPAYER_IDENTIFICATION_NUM_NOT_NULL = "사업자 등록번호는 빈 값일 수 없습니다.";
 	public static final String TAXPAYER_IDENTIFICATION_NUM_LENGTH_INVALID = "사업자 등록번호는 10자리 수입니다.";
 	public static final String TAXPAYER_IDENTIFICATION_NUM_FORMAT_INVALID="사업자 등록번호 형식이 올바르지 않습니다.";
