+++ src/main/java/com/project/seller/dto/request/SellerSignUpReuqest.java
@@ -3,6 +3,7 @@ package com.project.seller.dto.request;
 import static com.project.member.MemberConstant.*;
 import static com.project.seller.SellerConstant.*;
 import static com.project.seller.domain.TaxpayerIdentificationNum.*;
+
 import com.project.member.domain.Password;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -16,21 +17,24 @@ public record SellerSignUpReuqest(
 	@NotBlank(message = EMAIL_NOT_NULL)
 	@Email(message = EMAIL_PATTERN_INVALID + " : ${validatedValue}")
 	String email,
+
 	@Schema(description = "비밀번호", example = "test123!@#$")
 	@NotNull(message = PASSWORD_NOT_NULL)
 	@Pattern(regexp = Password.PASSWORD_REGEX, message = PASSWORD_FORMAT_INVALID)
 	String password,
+
 	@Schema(description = "사업자번호", example = "111-11-11111")
 	@Pattern(regexp = TAXPAYER_IDENTIFICATION_NUM_REGEX,
-			 message = TAXPAYER_IDENTIFICATION_NUM_FORMAT_INVALID)
+		message = TAXPAYER_IDENTIFICATION_NUM_FORMAT_INVALID)
 	@NotNull(message = TAXPAYER_IDENTIFICATION_NUM_NOT_NULL)
 	String TaxpayerIdentificationNum,
+
 	@Schema(description = "업체명", example = "하루야채")
 	@NotNull(message = COMPANY_NAME_NOT_NULL)
 	String CompanyName,
-	//TODO 업체 주소도 따로 포맷을 갖게 해야 할까?
+
 	@Schema(description = "업체주소", example = "서울특별시 영등포구 1111길 하루야채")
 	@NotNull(message = BUISINESS_LOCATION_NOT_NULL)
 	String BuisinessLocation
-){
+) {
 }
